Get dependencies on fit, slim and server straight.
