+++ ivy.xml
@@ -14,6 +14,7 @@
 		<conf name="default" visibility="public" />
 		<conf name="standalone" visibility="public" extends="default" description="include in fitnesse-standalone build" />
 		<conf name="test" visibility="public" extends="default" />
+		<conf name="slim" visibility="private" />
 		<conf name="maven" visibility="private" />
 		<conf name="lesscss" visibility="private"  />
 	</configurations>
@@ -27,7 +28,7 @@
 	</publications>
  
 	<dependencies>
-		<dependency org="org.htmlparser" name="htmlparser" rev="2.1" conf="default,maven->default">
+		<dependency org="org.htmlparser" name="htmlparser" rev="2.1" conf="default,slim,maven->default">
 			<exclude org="com.sun" module="tools" />
 		</dependency>
 		<dependency org="org.apache.velocity" name="velocity" rev="1.7" conf="default,maven->default" />
@@ -36,9 +37,9 @@
 		<dependency org="org.apache.ant" name="ant" rev="1.7.1" conf="default" transitive="false" />
 		<dependency org="com.googlecode.java-diff-utils" name="diffutils" rev="1.2.1" conf="default,maven->default"/>
 		<dependency org="org.eclipse.jgit" name="org.eclipse.jgit" rev="2.3.1.201302201838-r" conf="default,maven->default" transitive="false"/>
+		<dependency org="junit" name="junit-dep" rev="4.11" conf="default,test->default" />
 		<!-- scope: test -->
 		<dependency org="org.mockito" name="mockito-core" rev="1.9.5" conf="test->default" />
-		<dependency org="junit" name="junit-dep" rev="4.11" conf="test->default" />
 		<dependency org="org.hamcrest" name="hamcrest-all" rev="1.3" conf="test->default" />
 		<dependency org="net.javacrumbs.json-unit" name="json-unit" rev="1.1.6" conf="test->default" />
 		<dependency org="org.mozilla" name="rhino" rev="1.7R4" conf="lesscss->default" />
